+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Numeric => `close?`
 - Added Numeric => `delimit`
 - Added Numeric => `delta`
+- Added Numeric => `range`
 - Added String => `ascii_only`
 - Added String => `capitalized?`
 - Added String => `each_word`